Rename zip.{c,h} to archive.{c,h}
This should prevent confusion with the libzip, which provides the
`zip.h` header in `/usr/include` or similar.
